requirements.txt
asgiref==3.4.1
certifi==2020.12.5
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
ecdsa==0.17.0
idna==3.2
markdown2==2.4.0
numpy==1.20.1
oauthlib==3.1.1
Pillow==8.1.2
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pyo==1.0.3
python-dotenv==0.19.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
wxPython==4.1.0
